Enh - Initial visiting card webpage checkin.
This is a single page webpage based on startbootstrap-agency-1.1.0
The original template can be found in Gdrive -
 development > original_packages > startbootstrap-agency-1.1.0.zip

This was developed as a first point of access along with a visiting
card. The frontend directory in this checkin is hosted on Godaddy in
June, 2016.

This frontend will be replaced with a more suitable handcrafted
website in js+bootstrap+jquery , or with a reactjs in the near
future.

Checking this in now to document something that we hosted, and to make
progress on the in-house backend integration testing.

Tests
1. Ran the site locally on chrome and safari in various sizes.
2. Tested by running in production mode on Godaddy and tested on chrome
safari and iphone.