Enh - Integrating with signin/register/paswd modal
This is a hack integration of the signin modal with the
existing visiting card template.

The original modal code can be found in -
https://codyhouse.co/demo/login-signup-modal-window/index.html
and at -
development > original_packages > login-signup-modal-window.zip

There are a lot of issues with the conflicting css which needs
to be fixed.

Keep this as it is for now for integration with backend to go on.
This will be replaced with a handcrafed code or better integrated
code in the near future.