Update to Email templates
Updated email_base.html with Mailchimp template.
Updated content of the rest of the email templates.

Testing
- Tested in production mode .env of sending mail for
all the different flows.
- Used sendtestemail to test.

Todo
- The images in email_base.html currently reference the
mailchimp images. It needs to reference images stored in
frontend/img/email
- The password reset email does not display the link/button
to reset the email, needs to be fixed.
- the user_email_et.jpeg is the one edited in MC, needs to be
redone as the MC edit of the file is not good.
- Test via mailtrip.io